+++ src/App.css
@@ -10,6 +10,26 @@
   position: relative;
 }
 
+html, body, #root, #root > div {
+  padding: 0;
+  margin: 0;
+
+  width: 100%;
+  height: 100%;
+
+  display: inline-flex;
+  flex-direction: column;
+}
+
+main {
+  flex: 1;
+}
+
+footer{
+  background-color: grey;
+  flex: 0 0 10vh;
+}
+
 @keyframes App-logo-spin {
   from { transform: rotate(0deg); }
   to { transform: rotate(360deg); }
