+++ src/pages/Admin.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+class Admin extends React.Component {
+  render() {
+    return (
+      <div>
+        area de admin
+      </div>
+    );
+  }
+};
+
+export default Admin;
