+++ src/pages/Map.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+class Map extends React.Component {
+  render() {
+    return (
+      <div>
+        area de não admin
+      </div>
+    );
+  }
+};
+
+export default Map;
