+++ src/pages/Register.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+class Register extends React.Component {
+  render() {
+    return (
+      <div>
+        registrar
+      </div>
+    );
+  }
+};
+
+export default Register;
